fix(DTFS2-6886): fix API tests
