+++ portal/api-tests/login/post-check-your-email.api-test.js
@@ -1,0 +1,3 @@
+const { withSendNewSignInLinkApiTests } = require("./send-new-sign-in-link-api-tests");
+
+withSendNewSignInLinkApiTests('check-your-email');
