+++ portal/api-tests/login/post-sign-in-link-expired.api-test.js
@@ -1,0 +1,3 @@
+const { withSendNewSignInLinkApiTests } = require("./send-new-sign-in-link-api-tests");
+
+withSendNewSignInLinkApiTests('sign-in-link-expired');
